# PR: Reduce the duration of the AppVeyor tests and improve the performance of the gapfilling tool.