--- a/.gitignore
+++ b/.gitignore
@@ -64,6 +64,7 @@ Projects/
 .spyproject/
 @ new-prô'jèt!/
 
+Colors.db
 kgs_brf.exe
 WHAT.pref
 BRFInput.txt