The objective of this PR is to find out why tests are taking so long on AppVeyor and to fix the problem if possible.

The tests are taking more than 15 minutes on AppVeyor while they were taking less than 4 minutes on Travis. I suspect this has something to do with a `sleep` I've included in the gapfilling routine to prevent some glitch with the UI.

The test are passing after 3 min 23 sec on Travis.
The tests are now passing after 3 min 4 sec on AppVeyor.

This StackOverflow answer seems to explain pretty good what is probably happening here during the tests:
https://stackoverflow.com/a/25355498/4481445

When running the test on my machine:
Without the `sleep`, it takes around 17 seconds.
With the `sleep`, it takes around 125 seconds.

**So basically, this `sleep` call needs to be removed at all costs.** But, some testing are needed before I merge this into master.