@jnsebgosselin 2017-12-12
I've tested it with the UI and everything is working fine without the `sleep`. So this can be removed safely. An easy fix for a great gain in performance of the gapfilling routine.